DataFixtures, Paginator and some bugs fixed

+++ src/Controller/UserController.php
@@ -25,7 +25,7 @@ class UserController extends AbstractController
     /**
      * @Route("/", name="user_index", methods="GET")
      */
-    public function index(): Response
+    public function index() : Response
     {
         return $this->render('user/index.html.twig', ['users' => $this->userRepository->findAll()]);
     }
@@ -33,7 +33,7 @@ class UserController extends AbstractController
     /**
      * @Route("/new", name="user_new", methods="GET|POST")
      */
-    public function new(Request $request): Response
+    public function new(Request $request) : Response
     {
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
@@ -41,6 +41,10 @@ class UserController extends AbstractController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $em = $this->getDoctrine()->getManager();
+
+            $randomIdentifiant = $this->randomIdentifiant();
+
+            $user->setIdentifiant($randomIdentifiant);
             $em->persist($user);
             $em->flush();
 
@@ -56,17 +60,17 @@ class UserController extends AbstractController
     /**
      * @Route("/{identifiant}", name="user_show", methods="GET")
      */
-    public function show(User $user): Response
+    public function show(User $user) : Response
     {
         return $this->render('user/show.html.twig', [
             'user' => $user
-            ]);
+        ]);
     }
 
     /**
      * @Route("/{identifiant}/edit", name="user_edit", methods="GET|POST")
      */
-    public function edit(Request $request, User $user): Response
+    public function edit(Request $request, User $user) : Response
     {
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -86,15 +90,44 @@ class UserController extends AbstractController
     /**
      * @Route("/{id}", name="user_delete", methods="DELETE")
      */
-    public function delete(Request $request, User $user): Response
+    public function delete(Request $request, User $user) : Response
     {
-        if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+            if ($user->getNbBook() !== 0) {
+                $this->addFlash(
+                    'error',
+                    'L\'utilisateur ne peut pas être supprimé tant qu\'il possède un livre'
+                );
+
+                return $this->render('user/show.html.twig', [
+                    'user' => $user,
+                ]);
+            }
             $em = $this->getDoctrine()->getManager();
             $em->remove($user);
             $em->flush();
+
+            $this->addFlash(
+                'success',
+                'L\'utilisateur a bien été supprimé'
+            );
         }
+
 
         return $this->redirectToRoute('user_index');
     }
 
+    /**
+     * Generate random string for identifiant
+     *
+     * @return string
+     */
+    public function randomIdentifiant()
+    {
+        $randomIdentifiant = bin2hex(random_bytes(4));
+        if ($this->userRepository->checkIfRandomExists($randomIdentifiant)) {
+            $this->randomIdentifiant();
+        }
+        return $randomIdentifiant;
+    }
 }
